refactor: export common method for preparing tx for singining
